Compatibilidade do enum com swift

+++ AsAventurasDeMMM/AsAventurasDeMMM/GameScene.m
@@ -66,7 +66,7 @@
 }
 - (void)directionChanged:(AnalogGesture * )sender
 {
-    NSLog(@"%u",sender.analogDirection);
+    NSLog(@"%ld",(long)sender.analogDirection);
 }
 
 @end
